requirements.txt
asgiref==3.5.2
Django==4.0.8
django-countdowntimer-model==0.0.7
django-crispy-forms==1.14.0
django-debug-toolbar==3.7.0
django-guardian==2.4.0
django-multi-form-view==2.0.1
numpy==1.23.4
opencv-python==4.6.0.66
pandas==1.5.1
Pillow==9.3.0
python-dateutil==2.8.2
pytz==2022.6
six==1.16.0
sqlparse==0.4.3
tzdata==2022.6
